this commit finally was able to get the "cancel" button to kill the upgrade() process
A lot of things that I think *should* be working aren't, which could be due to this 7-year-old bug in kivy

 * https://github.com/kivy/kivy/issues/1116

Anyway, os.kill( pid, signal.SIGKILL ) does work (at least in Linux), but the pid is only accessible to multiprocessing.Process(), which doesn't make it easy to access the return value.

On the other hand, multiprocessing.Pool() gives easy access to the return value, but you can't get the pid of the processes in the pool.

Tomorrow I'll probably try to move this multiprocessing bits back from buskill.py to buskill_gui.py, if possible. And see if I can somehow extract the return value using multiprocessing.Queue() or Pipe() or something.