I finally converted buskill.py into a python class
I'm not sure it makes total sense to use objects for this, since it doesn't ever make sense for a client to ever create more than one BusKill object.

I primarily did this so that I can store the upgrade_status_msg in an instance field for passing data back-and-forth between the buskill_gui and the backend buskill.upgrade(), but--as this commit demonstrates--that is *still* not possible, even with BusKill as an object.

In fact, even after the process ends (and has been .join()ed), the instance fields that were updated within the amped-off process appear to be lost. Hmm

I suspect because of upgrade() is run in a distinct process, the instance of the BusKill object is copied in memory to a new/separate instance, and any merging of data between them would have to be done manually.

I should do more research on the right way to do this. How do Kivy developers execute stuff in the background so the UI doesn't freeze, but their object's long-running methods can still update their instance fields in an accessible way?