added shared memory to get output from upgrade()
I implemented a proper multiprocessing.Array() so that there's now shared memory (a ctypes character array = string) to pass between the parent & child processes.

This commit uses two of these arrays to successfully get the following from the upgrade() function run in a child process launched with multiprocessing.Process():

 1. The return value (which is either '1' if the user is already running the latest version or a string with the absolute path to the new executable that was downloaded)

 2. The upgrade_status_msg that I've been trying to get working!

So this commit now actually updates the GUI dialog with [2], so the user can see the progress of the upgrade() as it runs. Awesome!

Note that I implemented this in such a way that the user can still all upgrade() directly and it will still return integers or strings like normal without needing to use shared memory. That whole setup is only required if upgrade_bg() is called (which sets-up the shared memory and executes upgrade() in a child process using multiprocessing.Process())

Unfortunately, what this commit does *not* solve is exception catching. If I were using multiprocessing.Pool(), then I could call .get() and catch exceptions trivially (and also get the return value of the function without having to create the shared memory Array for [1]) -- but then I would have to add another shared memory anyway to get the pid so I could terminate it as needed.

Or I could just subclass multiprocessing.Process() with something that overrides run() and handles stores exceptions?

 * https://stackoverflow.com/a/33599967/1174102