exception fighting
this time I got this error message in the modal:

   exceptions must derive from BaseExeption

I'm flying blind, but I'm pretty sure this is from the get_upgrade_result() trying to 'throw' a string. And I think the previous erorr originated from here as well. If I keep fighting this, I'll just do a throw "RuntimeWarning( exception )".

It would actually loose the real exception, but keep the text and work.