attempting to override Process() without the try/exception logic
The last build appears to go all the way through the upgrade process. Then at the very end, it opens this error in the "Update Failed!" modal:

  'Process' object has no attribute 'exception'

So (because the exception catching logic has been removed), it's not clear if there was an exception or not, but it did appear to make it at least most of the way through the upgrade() process

This commit attempts to do the override of the multiprocessing.Process() method, but with what I think is the exact same contents. If the behaviour is the same as before, then I know for sure that the issue is in that try/except block