attempting to trigger pickle error
I fixed the new pip install issues on the last build.

Somehow the last build worked perfectly..it even caught an exception and displayed it in the error modal dialog perfectly at the end of the install when it was extracting where a file already existed.

This commit adds a 'raise' to try to trigger the pickle error