attempting to define a __reduce__() method for the BusKill class
Since I've obliterated all of my changes to override multiprocessing.Process() and it's *still* having pickle errors, I think the issue may be that I changed upgrade_process to be an instance field (self.upgrade_process) so now the BusKill object itself contains a variable that is the return of Process(), which itself might not be pickleable--causing this issue.

This commit defines a __reduce__() method that returns a dumb string (a bogus serialization). If the error doesn't exist, then I know that I've pin-pointed the issue and I can actually define the __reduce__() function properly to exclude non-serializable instance fields. Or, maybe, use the `dill` module to pickle the unpickleable