trying to use __getstate__() to just omit instance fields that cannot be pickled
using __reduce__() worked great for us to pin-point the issue generally, but now that the child process is given a fresh object, it doesn't have access to our shared memory = self.upgrade_status_msg & self.upgrade_result.

The consequence of this is that previous builds can't update the UI with the status of the upgrade() or get the path to the extracted upgrade after installing it.

This commit attempts to pin-point the issue more directly by actually sending the instance of our BusKill object off to the child process, but by removing the instance fields that are not pickleable. My hope is that the two above-listed multiprocessing.Array() fields *are* pickleable, but only the self.upgrade_process object can't be pickled. If this doesn't work, then in future commits I'll just keep removing items from the state dictionary until I've pin-pointed exactly what instance fields in our BusKill object cannot be pickled.