removing 'upgrade_status_msg' from the unpickleable list to see if will cause the pickle exceptions to return
The last build worked perfectly as in I couldn't get the pickle errors to return. Instead, this exception is caught from the child process and correctly displayed on the modal UI dialog:

 'BusKill' object has no attribute 'upgrade_status_msg'

..which is expected because a new initialized BusKill instance will have upgrade_status_msg == None.

This commit attempts to remove that instance field from the unpickleable list in __getstate__(). If the pickle errors return, then I think the mulitprocessing.Array() variables are not pickleable