removing our other Array() from the unpickleable list
The last build worked beautifully. No more pickling or SynchronizedString or pickle errors. This suggests that the multiprocessing.Array() instances can, in fact, be pickled. That's great!

This commit removes our second Array() from the unpickleable list so we can actually get the upgrade() result in addition to the status message.

I expect that everything should work on MacOS after this, up until the "Restart Now" button is pressed. Wow that was an annoying rabbit hole..