hard-coding the size of the progresspinner to '70'
The last build that now has a more robust auto-sizing of the modal dialog's height for the body text means that, as the upgrade_status_msg changes, the size of the progress spinner changes.

Unfortunately, I couldn't get the progresspinner to accept a height, width, or size. It appears to ignore all my inputs. For now, this commit just hard-codes the size of the progress spinner to be '70'