fixing window resizing
I suddenly realized that the window couldn't be resized. Looks like the root-cause was a logic error in the super() call inside the __init__() function I put in the MainWindow() class that overrides BoxLayout().

The class in that super() call should be the current class (MainWindow)--not the super of that class (BoxLayout). What I had was like calling the super on the super, which is not what I want..