adding upgraded_to.py so an old version knows that it's been replaced if a newer version has been installed via the app's auto-update tool
To achieve this, I'm writing-out a python file named 'upgraded_to.py' in the same dir where the executable lives (EXE_DIR) and checking for it [a] at startup and [b] when the user clicks the "Update" button in the GUI.

In either case, if the updated_to.py file (which contains a dict specifying the path to the new executable that deprecated the current version [EXE_PATH]) points to a replacement executable that exists, then the user is prompted to restart to the new version instead.

To reach this end, I've also standardized some of the variables that I was using before (chaotically spread across distinct functions and scripts) into instance fields of the BusKill object:

```
 NOTE about instance fields used for storing path info relative to the
      buskill executable:

 1. EXE_PATH = the absolute path to the executable for running BusKill,
               which is:
                 [a] the 'buskill-<version>.AppImage' file in Linux
                 [b] the 'buskill-<version>.exe' file in Windows
                 [c] the 'buskill' binary file in MacOS
 2. EXE_FILE = just the filename (basename) of the EXE_PATH
 3. EXE_DIR = the directory (dirname) where EXE_FILE lives
 4. APP_DIR = the root directory for a given version of the buskill app,
              which is 'buskill-<lin|win|mac>-<version>-x86_64' and:
                 [a] the dir containing the EXE_FILE file in Linux
                 [b] the dir containing the EXE_FILE file in Windows
                 [c] 2 dirs above the dir containing the EXE_FILE in
                     MacOS
 5. APPS_DIR = the directory where the app dirs live (one dir above
               APP_DIR)
```

Because of the above changes, this commit may break the upgrade process on Windows & MacOS due to off-by-one path dir depth or something.

Anyway, that may need to change already because:

 1. Turns out my windows .zip release is a "zip bomb". In Windows it appears to extract into a dir by default, but on Linux it just dumps all the files into cwd. I should change this so the root of the .zip is a single dir named 'buskill-win-v0.1.0-x86_64'

 2. I realized that I should be putting a few other basic files into my release dirs, including (at least) the LICENSE, attribution, and a docs/ dir with a simple readme that just directs the user to visit docs.buskill.in

Next I'll be finishing 'updated_from.py' -- which is a file that's (unlike 'upgraded_to.py') placed into the *new* version's APP_DIR so that when the new version is executed, it can delete the old version's APP_DIR (and the updated_from.py file itself_ on its first start.