removing " | Out-String" from env assignments
The last build proved that appending " | Out-String" apparently adds a newline to $tmpDir, which makes it break when I pass $tmpDir to the `mkdir` commands, so this commit removes that bit.

The " | Out-String" was added to force PowerShell to execute each line sequentially and synchronously, which apparently it doesn't do (at least in GitHub's shared runners) by default. Hopefully removing it won't result in "variable referenced before assignment" issues. If so, then I'll have to add it back and figure out some way to strip() off newlines at the end.