cleanup debugging changes
actually, it appears that the reason why I was getting integrity issues over-and-over was just because my internet is shitty. After a particularly bad period of packet loss passed, everything worked perfectly

So this build should be fully functional on linux and windows. Unfortunately, I don't have physical access to a Mac to test the arm/trigger functionality on Mac. But if it works on Linux then it's likely to work on Mac, since both use libusb (Windows is the weird one that uses CPython and some Microsoft stuff)