getting more info about GitHub shared runner's network/processes during run
So the last build stalled indefinitely and ran for 48 minutes before exiting. Clearly, the instance itself needs to be able to communicate out to GitHub's servers in order to finish. And I appear to have broken that with my restrictive iptables settings.

This commit attempts to print enough info for me to figure out what the minimum holes I need to poke in the firewall are such that I don't break the underlying communication out from the GitHub shared runner instance back to GitHub's control servers

See also:

 * https://stackoverflow.com/questions/64090733/iptables-rules-to-permit-github-actions-from-breaking-chains-default-to-drop/64091585#64091585