adding debian wrapper script to execute linux build script in debian:stable-slim docker container
I discovered recently that Docker Content Trust actually provides zero security for our setup. It's 100% security theater since it blindly TOFUs the root key on every build (without so much as even printing an INFO message, and it's not even clear what's happening in the debug output of `docker pull`)

 * https://github.com/BusKill/buskill-app/issues/6#issuecomment-700050760
 * https://security.stackexchange.com/questions/238529/how-to-list-all-of-the-known-root-keys-in-docker-docker-content-trust

This commit updates our CI process to execute a new script: debianWrapper.sh, which executes our linux build script inside a debian docker container so that it's reproducible *and* has the NET_ADMIN capability as described in the issue linked-to above.

We make the download of the debian docker image actually secure by pinning the root public key used by the debian team to sign their docker images, which are added with this commit and copied in-place to the /root/.docker/trust dir by the debianWrapper.sh script before calling the `docker pull` command.

Moreover, we check to confirm that the cryptic line "200 when retrieving metadata for root" does *not* appear in the debug output of the `docker pull` command, which--afaict--is the only way to know if docker is dumbly downloading/tofu-ing the root key used to sign the docker image. AFAIK, There is currently no other way to prevent `docker pull` from doing this. See bug report to change this here:

 * https://github.com/docker/cli/issues/2752