making the wrapper run as root
This commit changes the debian wrapper script (which needs root to be able to do some docker-host-setup stuff) as root. If it's not already running as root, it'l call itself using sudo to make sure it runs as root. And the CI workflow is now changed to run this script using sudo so it runs as root in the GitHub shared runner.