polish
the last build worked at making the .app have the icon in Finder (both after "installl" and when viewed in the dmg), but the dmg still doesn't have the icon. And, of course, the icon still works in the taskbar when buskill-app is running.

The dmg is short-lived, so I don't consider that essential. I left a TODO commment in the dmg build script, and I'm going to move-on to Windows icon tests