fix regex before deleting old version to work on both *nix and windows
The last windows build mostly worked, but--after restarting--the new version of the app dies with an error

  re.error: unterminated character set at position 10

This is because of some confusion with backslashes as the escape character and the os.sep character in Windows. This commit adds a hack that appears to work for both *nix and Windows.