moving buskill.log outside the APP_DIR
The last build successfully implements the in-app upgrade functionality from v0.2.1 to v0.3.1 on Linux and MacOS, but it fails on Windows with the error

  	The process cannot access the file because it is being used by another process: 'C:\\Users\\user\\Downloads\\buskill-win-0.2.1-x86_64\\buskill.log'

This occurs when attempting to `shutil.rmtree()` the old `self.UPGRADED_FROM['APPDIR']`. The root cause is that, for some reason, the restart after upgrade doesn't change the log dir, so the old buskill.log file is still open, causing the exception and crash of the app.

This commit fixes this by moving the buskill.log file the temp dir. I only had it in the APP_DIR on Windows because I wasn't sure where the Windows temp dir was. Well, in this commit I'm using the cross-platform, built-in tempfile.gettempdir() to figure that out.