catching errors trying to delete old version after upgrade
The last build still work on Linux & MacOS, but for some reason I'm still getting a PermissionError on Windows:

  PermissionError: [WinError 32] The process cannot access the file because the file is being used by another process: 'C:\\Users\\user\\Downloads\\buskill-win-v0.2.1-x86_64\\buskill-win-v0.2.1-x86_64'

In fact, after the app crashes, I discovered that it *did* delete all of the contens of the above dir, but the dir itself didn't get deleted. When I ran v0.3.1 manually, it finished the deletion.

Anyway, I guess this is some weirdness with how execv() is run in Windows. So this commit just wraps the deletion in a try/except statement and issues a warning if the deletion fails, rather than exit on the uncaught exception. I expect that it will WARN on the first run, then it'll finish off the deletion of the empty dir the next time the new version is started.