platform-agnostic slashes for APP_DIR
The last build still has issues after upgrade() on Windows. The app restarts and catches the deletion exception error. But then when the app loads, the debug log gets spit-out wiht a ton of errors. And the app, oddly, has no text. I guess it's because, maybe, it is trying to use the font files from the old version? Winodws execv() is weird.

Anyway, I did notice that the APP_DIR is empty because I was using the forward slash (/) as the path delimiter on Winodws, which is wrong.

This commit changes that code to use os.sep, which is the platform-agnostic solution. Maybe this will fix the restart error? I doubt it, but it needs to be fixed, anyway.