attempting to cleanup the env by removing the old version before the restart after upgrade()
While the recent commit to load the fonts at an absolute path worked for Windows, it broke Linux because the linux AppImage actually extracts itself at runtime to a temp dir. This temp dir puts the resources (like fonts) in a distinct path from EXE_DIR (which is the dir where the AppImage itself lives). One option is to define a new buskill instance field to be self.RESOURCES_DIR or something. Or we just go back to the old code, which I guess uses cwd or searches PATH. Not sure.

Anyway, this commit attempts to fix the underlying restart issue on Windows by instead cleaning-up the env before the restart--removing all references to the old version. We'll see if it works..