adding prompt for manual restart on Windows
The last build entirely got rid of all the environment variables that were referencing the old version, but it *still* freezes after restart (undeterministically).

Anyway, I'm just giving up and going to try to prompt the user to do a manual restart if/after the recursive delete of the old version fails to finish the deed. Hopefully *this* won't freeze..