attempting to remove command prompt window from windows builds
this commit adds the --noconsole argument to the pyinstaller build of
windows buskill.exe builds. Our existing build opens both the GUI app
and also a command prompt terminal behind it in another window.

The con of this change is that we won't be able to use the app in CLI
mode on windows. Well, unless we do some hack like appending "| more"

 * https://stackoverflow.com/questions/38981547/making-a-pyinstaller-exe-do-both-command-line-and-windowed

If this works, then I should consider opening a feature request with
PyInstaller and do some testing to find the most elequant solution
(long-term, maybe that means having a distinct cli-only build -- which
would also be insanely smaller without all the kivy depends)