better error detection with $? instead of $LastExitCode
TIL that $LastExitCode in powershell will not be updated if the
previously-executed command fails because the command is not found.
Therefore, it will prevent our Windows build script from exiting when
`gpgv` fails because `gpgv.exe` cannot be found.

This commit fixes that by using $?, which does work as expected