adding scripts to build buskill's usb drive root
this is the first commit to address issue #22

 * https://github.com/BusKill/buskill-app/issues/22

The result of executing `build/usb/debianWrapper.sh` is a 7zip archive that (when unzipped) spits out a `usbRoot/` directory. The contents of that `usbRoot/` directory should be copied straight into the root of a freshly-formatted fat32 usb drive.

This USB drive's contents will be cross-platform but also intended to be as easy-as-possible for the user to launch the BusKill app.