fix reST processing error
this commit is attempting to fix this bug where the featured image and the toc on the main page somehow get pulled inside the notebox above it, despite the lack of indentation.

it renders fine locally, but for some reason it's broken on GitHub. Let's see if this fixes it.