updating the KEYS file with my key after it's been extended
I'm working on the script to generate the contents of the USB drives
that ship with the BusKill cable kit

 * https://github.com/BusKill/buskill-app/issues/22
 * https://github.com/BusKill/buskill-app/blob/dev/build/usb/debianWrapper.sh

One thing I explicitly add to the usb archive (in addition to the
obvious software for each platform) is the source archives for each
platform and their cryptographic signature. And I include the keys used
to sign it. This is stored in the 'sigs' dir. Obviously, if the contents
of the USB drive is maliciously alterted, then this is not itself
trustworthy. But if it is modified, then it'll be very easy to detect
because my public key fingerprint can be found on a bunch of distinct
web servers and domains..

For future reference, I updated this with the following commands:

tmpDir=`mktemp -d`
pushd "${tmpDir}"
echo "" > KEYS

  gpg --list-keys --armor 'E0AF FF57 DC00 FBE0 5635  8761 4AE2 1E19 36CE 786A' >> KEYS
  gpg --list-sigs 'E0AF FF57 DC00 FBE0 5635  8761 4AE2 1E19 36CE 786A' >> KEYS
  gpg --export --armor 'E0AF FF57 DC00 FBE0 5635  8761 4AE2 1E19 36CE 786A' >> KEYS

  gpg --list-keys --armor '713D 4A49 60EE 849B AE3B  41BA BE75 DB07 E34A FBC1' >> KEYS
  gpg --list-sigs '713D 4A49 60EE 849B AE3B  41BA BE75 DB07 E34A FBC1' >> KEYS
  gpg --export --armor '713D 4A49 60EE 849B AE3B  41BA BE75 DB07 E34A FBC1' >> KEYS

  gpg --list-keys --armor '0465 E42F 7120 6785 E972  644C FE1B 8449 4E64 0D41' >> KEYS
  gpg --list-sigs '0465 E42F 7120 6785 E972  644C FE1B 8449 4E64 0D41' >> KEYS
  gpg --export --armor '0465 E42F 7120 6785 E972  644C FE1B 8449 4E64 0D41' >> KEYS

Which is roughly documneted here:

 * https://docs.buskill.in/buskill-app/en/stable/security/pgpkeys.html#developers