attempting to fix PyInstaller
not sure why I ever attempted to upgrade PyInstaller via pip. That's not
secure. And I noticed it breaks builds since January 2022. Later builds
appear to take the pinned PyInstaller v3.6 and upgrade it with a
PyInstaller 5.0.1 -- which is breaking the builds.

Let's try to comment-out ths upgrade line and see if the builds get
fixed