change to use screen manager
this commit actually changes the root kivy widget from BoxLayout to ScreenManager. I added a (global) instance field name 'sm' (ScrenManager) to the BusKillApp() class, which is inhereted and set by both the MainWindow() and BugReport() classes. This allows them to call the screen manager and switch back.

Probably there's better ways to do this than global variables, but it works..