improvedments to TextInput loading the Debug Log
From a few commits ago:

1. I fixed it to load async in a thread. This means the UI doesn't lock-up anymore when it loads the debug log, which could take up to 20 seconds if the file is 3 MB+

 * https://stackoverflow.com/a/72928947/1174102

2. The user still can't select-all with ctrl+a. This was actually a bug that was fixed in Kivy v2.1.0. I tested it in the newer version and confirmed it's fixed

 * https://github.com/BusKill/buskill-app/issues/29#issuecomment-1179709462

Unfortuantely, while the async is certainly better, if the user resizes the window then Kivy *really* struggles to adjust the contents of the TextInput. It basically freezes the UI. Even for a debug log that's <0.5 MB. This is really bad.

Another option is to give-up on TextInput and just go with a RecycleView of labels with one label per line. The user won't be able to select the text at all, but it'll be fast. Anyway, I already have a button for "Copy Log" so it's not strictly necessary for them to be able to select bits of text from the log shown on the screen.