switched TextInput to RecycleView with one-line labels
This commit gives-up on TextInput and switches instead to a RecycleView

 * https://stackoverflow.com/a/72923330/1174102

This is *wayyy* more efficient for kivy to render the text. It resizes without locking-up, and the data loads almost instantly.

Problems:

1. The user can't select the text
2. I lost all my formatting and I can't quite figure out how to change that (font mono, font size, align-left, etc)
3. I got the scroll view to work for both "content" (so the user can click/drag around on the text itself) and "bars" (so the user has a scrollbar on the right-hand-side, which is absolutely essential to move fast acros a huge log file's contens), but things like "page down" and "end" don't move around in the ScrollView as you'd expect

But it took me a bit to even get this functional, so I'm comitting now while it works. I'll iron-out the details in future commits.