I'm giving-up on getting the width of the RecycleView scroll box to match the longest line. I'm just going to leave it as a TODO in the comments and move-on. For now I'm just making it 3000 pixels wide, which appeares to be more than enough to fit what's in my debug log now
There is one exception: there is one line (the print of os.environ) which is so long, it appears as a single black line. IIRC this is an OpenGL overflow bug or something. Anyway, this view of the log file in the app isn't especially importnat. What's important is that the actual text gets copied to the clipboard for submitting a bug. And that's working. So I consider this black line to be a very minor bug.

Moving on.

I'd next like to display a tooltip next to the label at the top above the debug log, which tells the user where the log file is stored -- so they can just open the damn thing in a proper text editor and not be limited by kivy in viewing the log, if they so desire