downgradeing docker build from Debian 11 (bullseye) to Debian 10 (buster) since it appears to cause the AppImage to fail to build
 * https://github.com/BusKill/buskill-app/releases/tag/2650735006_linux
 * https://github.com/BusKill/buskill-app/runs/7285758444?check_suite_focus=true

This could probably be fixed. iirc I'm building the mksquashfs binaries from source because it was needed to get fully reproducable builds for the .AppImage, and at the time tme mksquashfs binary on Debian broke reproducability.

So if I update the python AppImage and the AppImageTool itself (and maybe PyInstaller, which has made huge strides in supporting reproducable builds in the past years), then this may go away.

But for now, I'm just going to build on Debian 10.

	2022-07-11T15:43:53.9664844Z + /tmp/appimagetool_appdir/AppRun --no-appstream /tmp/kivy_appdir dist/buskill-lin-gui-debug-log-x86_64/buskill-gui-debug-log.AppImage
	2022-07-11T15:43:53.9718397Z appimagetool, continuous build (commit effcebc), build 2084 built on 2019-05-01 21:02:41 UTC
	2022-07-11T15:43:53.9918813Z Using architecture x86_64
	2022-07-11T15:43:53.9919073Z Generating squashfs...
	2022-07-11T15:43:53.9947597Z /tmp/appimagetool_appdir/usr/bin/../lib/appimagekit/mksquashfs: 8: /tmp/squashfs4.4/squashfs-tools/mksquashfs: not found
	2022-07-11T15:43:53.9952934Z Embedding ELF...
	2022-07-11T15:43:53.9953484Z /tmp/kivy_appdir should be packaged as dist/buskill-lin-gui-debug-log-x86_64/buskill-gui-debug-log.AppImage
	2022-07-11T15:43:53.9954176Z Not able to open the AppImage for writing, aborting

The long-term