Attempt to split buskill/__init__.py
this commit is my first attempt to split-up the contents of the buskill __init__.py file into distinct files. Here I've only attempted to make upgrade() separate, following this:

 * https://www.qtrac.eu/pyclassmulti.html
 * https://stackoverflow.com/a/57562083

Note that the new builds could't find the KEYS file (which here not in the APP_DIR on Linux, but in the root-owned /tmp/ dir in which the squashfs .AppDir is extracted), so I had to do some fuckery with APP_DATA and added a new instance field RUNTIME_DIR

I've tested that this works on Linux, but I'm not sure if this works on Windows & MacOS (hence why I'm pushing this to git for the builds).

If I broke the upgrade process on any platforms, I'll revert this

 * https://github.com/BusKill/buskill-app/issues/33