attempt to get the paths working for executing a child process as root (needed for the soft-shutdown trigger)
 * https://github.com/BusKill/buskill-app/issues/14

This commit adds an instance filed named "SRC_DIR" to the buskill object, which is where we place files like KEYS and also where the 'packages' dir lives. This is necessary so we can specify the aboslute path to the 'packages/buskill/root_child_mac.py' file. Note that this SRC_DIR is just the EXE_DIR (the dir that holds the buskill executable) on Windows and MacOS, but it's actually in a root-owned temp dir on Linux (where the AppImage was extracted-to at runtime)