attempt to execute mock root_child_mac.py when setting the trigger to 'soft-shutdown'
this commit updates the build script to (try to) give the root_child_mac.py script root:root 0400 permissions.

And it updates the set_trigger() to call spawn_root() when the trigger is set to 'soft-shutdown' (since soft-shutdown on MacOS requires root permissions). At this point, it just verifies the sanity of the root_child_mac.py script (for security) and then attempts to launch it as root, but the script currenly only tries to write-out a file to a root-owned directory as a POC.

I need to test this build, there will probably be errors. Once I can actually launch the script as root properly, then I'll fill-out the script with actual logic to accept commands and, if passed the 'soft-shutdown' command, logic to shutdown the machine

 * https://github.com/BusKill/buskill-app/issues/14