fixed --run-trigger functionality to be more realistic
The previous implentation of --run-trigger just called `bk.TRIGGER_FUNCTION()` directly. This was a problem, however, because a *real* trigger would execute the callback function running inside a non-root child process. So testing via --run-trigger wouldn't work for simulating, for example, the non-root child process trying to communicate to the root child process for executing the soft-shutdown trigger on MacOS.

This commit changes the functionality to call the callback function, just as would happen with a *real* hotplug removal event.

 * https://github.com/BusKill/buskill-app/issues/14#issuecomment-1281714319
 * https://github.com/BusKill/buskill-app/issues/38