attempt to create a new logger in root_child_mac that writes to the same file as the parent process
 * https://github.com/BusKill/buskill-app/issues/14#issuecomment-1284569270

In my tests, I found that I cannot inheret the logger from one process to another. However, I did find that I could write to the same log file from multiple processes (and it's supposed to be thread-safe to do so).

I tested manaully defining the log file, and it worked. The next question is how to pass the log file from parent process to root child.

I really don't like the idea of passing any *input* into the process running as root, but I'm afriad we can't see what the hell it's doing unless it's logging. I could have it write to its own log file. Or to *guess* where to log to, but then it would make troubleshooting user bug reports extremely difficult. It's actually really important to make it easy for the user to submit a single debug log with everything we need to understand what's going on.

So, I decided to pass the log file as the first positional argument to the root_child_mac on execution. I do a pretty strict regex match on it before accepting it, and exit if it has anything other than alpha numerics, spaces, periods, dashes, and underscores in it.

Hopefully this works.