updates the background of the action bar of all screens to have the same color as the main screen.
This commit improves the change from two commits ago that fixed the Debug Log (and new Settings) screens to have a "blue" instead of a "black" background color to meet the Material Design Spec.

The problem was that we actually don't want to be blue all the time. If the app is armed, then the background was red -- but only on the Main Screen.

Now we loop through all the screens (and loop through all the Action Bar's children on all the screens) to update all our app's Action Views to switch between red and blue on arm/disarm