Removed "Save" button on "Settings" screen
The material design spec doesn't mention anything about a save button, but after poking around Android settings, I realized that the settings are actually applied automatically. There are no save buttons

 * https://m1.material.io/patterns/settings.html#settings-usage

After reading this UX question on SE, I came to the general conclusion that if a user types some input that needs validation, then maybe a save button is warranted. But if it's just choosing a toggle or checkbox, then I don't think it makes sense. If anything, where needed, I'll just present the user with a modal pop-up asking them to confirm at the time they click it.

 * https://ux.stackexchange.com/questions/12289/guidelines-for-when-to-auto-save-settings-vs-save-button

For example:

 * if the user sets the trigger to '*-shutdown', I should warn them about the risk of data loss and make them confirm before the setting is applied

 * if the user changes the trigger while BusKill is armed, then I'll present them with a pop-up telling them that the current trigger is set to '<old-trigger>' but they've now changed to '<now-trigger>' and ask them if they wish to re-arm BBusKill with the '<new-trigger>'