solved a way to reference the App from a Screen
This commit isn't pretty, but I'm going ahead and commiting now because I found a way to actually reference the App from within a Screen object, which was the problem of the most-recent commits that prevented me from getting rid of the global 'bk' variable.

I actually didn't solve this because of the 'global' variable, but I wanted to be able to call open_settings() from within a Screen -- which was the same problem as before. I needed to call app.open_settings() from within a screen, which couldn't find the App.

I ended-up solvin this by adding an on_start() function tto the App that iterates through all the screens in the manager and adds an instance field named 'root_app' to each screen object in 'manager' and sets that to the BusKillApp's 'self'

Then I proved that I was able to access this in the "Settings" object (which is a Screen) when a button is clicked.

However, I was *not* able to access this 'root_app' instance field from within the MainWindow on_start() function, perhaps because it's called too early before it's set. I need to fix this.