half-fixes reference to BusKillApp from Settings Screen
This commit is successfully able to reference the 'bk' instance field in the main BusKillApp object, but only after you change screens.

For example, if you click the "arm" button immediately after opening, the app will crash because it cannot find the 'bk' object

   File "/home/user/sandbox/buskill-app/src/buskill_gui.py", line 108, in toggle_buskill
     self.bk.toggle()
 AttributeError: 'MainWindow' object has no attribute 'bk'

This occurs because I was only able to set the object with on_leave().

Because I use on_leave(), the app only doesn't crash on the "Arm" button is pressed *after* changing the screen from MainWindows. So, for example, if you open the Setting scree, then tap "back", then you can hit the "Arm" button without it crashing, because the 'root_app' instance filed is created a the time you left the MainWindow Screen to open the Settings Screen

I'm only doing on_leave() as a proof-of-concept because I still get crashes if I try to put the logic that instantiates self.bk in on_enter() and on_pre_enter() saying that 'root_app' doesn't exist yet. So I need to figure out a way to set 'root_app' from within BusKillApp first, and then set 'self.bk' to 'root_app.bk' from within MainWindow before the user has a chance to interact with the UI