fixed eliminating 'global bk'
This commit solved the issue of refernecing the App from within the Screen by separating the instantiaion of each Screen from adding it to the widget.

Now we create teach screen, then we add the 'root_app' instance field to each screen, and *then* we add the screen as a widget to the screen manager.

This allows us to access the 'root_app' instance field within each instance of a Screen's on_pre_enter() funtion (but not __init__(); that's too early), so we can setup the screen's 'self.bk' object by setting it to 'self.root_app.bk'

TODO: cleanup. this was just a fast commit after I finally got it working!