got settings to appear without double-ActionBar
I've been fighting for the past few hours because, for some reason, creating the Settings widget had its own copy of my ActionBar. Maybe I'm using it wrong, but the examples on how to use it are pretty slim. Anyway, no matter how I tried to add the settings widget to my Settings window, it kept adding it in a narrow sliver on the right-half of the window. On the left-half was a redundant ActionBar.

This code is sloppy, but I'm commiting now because I finally got it to work. All I'm doing is iterating through the children of the Settings widget and removing the widget that's of type = BoxLayout, because that contains the ActionBar, apparently.

I don't know why this happens or if there's a better solution, but it finally works.