prevent settings widget from being added over-and-over
The previous commit had a bug where we kept adding the settings widget every time the screen changed to the Settings screen. This is because we implemented the code to add the settings widget to the settings_content BoxLayout on_enter().

We added the logic on_enter() because we can't add it on __init__() because we can't yet access 'root_app' at that stage (it's too early).

Anyway, this commit fixes the issue just by checking the children of the 'settigns_content' BoxLayout. If it's empty, we add it. If it's not empty, we don't add it. It works.